D0915  FIN Time: 4:04 P.M.

+++ CYUTGP110IME/View/ForumView.swift
@@ -7,7 +7,8 @@
 
 import SwiftUI
 
-struct ForumView: View {
+struct ForumView: View
+{
     //存取登入狀態
     @AppStorage("logIn") private var logIn: Bool=false
     
@@ -26,11 +27,13 @@ struct ForumView: View {
     @State private var forum: [Forum]=[]
     
     //MARK: 刪除文章
-    private func deleteForum(id: String) async {
+    private func deleteForum(id: String) async
+    {
         FireStore().deleteData(id: id)
     }
     //MARK: 刪除使用者
-    private func deleteUser() async {
+    private func deleteUser() async
+    {
         //Authentication
         Authentication().delete()
         //Realtime Database
@@ -38,13 +41,16 @@ struct ForumView: View {
         //User結構
         self.user.deleteUser()
         //刪除之後登出
-        withAnimation(.easeInOut.speed(2)) {
+        withAnimation(.easeInOut.speed(2))
+        {
             self.logIn=false
         }
     }
     //MARK: 查詢文章
-    private func getForum() async {
-        withAnimation(.easeInOut) {
+    private func getForum() async
+    {
+        withAnimation(.easeInOut)
+        {
             //顯示ProgressView
             self.progress=true
             //初始化圖片避免重複
@@ -57,18 +63,22 @@ struct ForumView: View {
         //查詢Firestore Database的文章資料
         FireStore().getData {(success, error) in
             //有查詢結果
-            if let success=success {
+            if let success=success
+            {
                 //遍歷每一個資料欄位
-                for i in success.documents {
+                for i in success.documents
+                {
                     //MARK: Storage
                     //下載對應文章的圖片
                     StoRage().downloadPicture(path: "forum", name: i.documentID) {(data, error) in
                         //將Data轉換成UIImage
                         if let data=data,
-                           let image=UIImage(data: data) {
+                           let image=UIImage(data: data)
+                        {
                             //將UIImage放進image
                             self.image.append(image)
-                        } else {
+                        } else
+                        {
                             //沒有圖片
                             self.image.append(nil)
                         }
@@ -87,13 +97,15 @@ struct ForumView: View {
                     }
                 }
                 //查詢錯誤
-            } else if let error=error {
+            } else if let error=error
+            {
                 self.result.1=error.localizedDescription
                 self.result.0.toggle()
             }
             
             //停止顯示ProgressView
-            withAnimation(.easeInOut){
+            withAnimation(.easeInOut)
+            {
                 self.progress=false
             }
         }
@@ -102,19 +114,23 @@ struct ForumView: View {
     var body: some View {
         ZStack {
             //MARK: ProgressView
-            if(self.progress) {
+            if(self.progress)
+            {
                 ProgressView().transition(.opacity)
                 //MARK: 預設畫面
-            } else if(self.forum.isEmpty && self.image.isEmpty) {
+            } else if(self.forum.isEmpty && self.image.isEmpty)
+            {
                 Text("Almost Done!").font(.largeTitle).transition(.opacity)
                 //MARK: 文章
             } else {
                 List {
                     ForEach(self.forum.indices, id: \.self) {index in
                         //MARK: 作者
-                        Section(self.forum[index].author) {
+                        Section(self.forum[index].author)
+                        {
                             VStack {
-                                if let image=self.image[index] {
+                                if let image=self.image[index]
+                                {
                                     //MARK: 圖片
                                     Image(uiImage: image)
                                         .resizable()
@@ -122,7 +138,8 @@ struct ForumView: View {
                                         .cornerRadius(10)
                                 }
                                 
-                                VStack(alignment: .leading, spacing: 0) {
+                                VStack(alignment: .leading, spacing: 0)
+                                {
                                     //MARK: 標題
                                     Text(self.forum[index].title)
                                         .bold()
@@ -148,15 +165,19 @@ struct ForumView: View {
                         .listRowBackground((self.forum[index].secure ? Color.red:Color.green).opacity(0.5))
                         .listRowInsets(EdgeInsets())
                         .swipeActions(edge: .trailing) {
-                            if(self.forum[index].author==self.user.name) {
+                            if(self.forum[index].author==self.user.name)
+                            {
                                 //MARK: 刪除文章Button
-                                Button {
-                                    Task {
+                                Button
+                                {
+                                    Task
+                                    {
                                         let id: String=self.forum[index].id
                                         await self.deleteForum(id: id)
                                         await self.getForum()
                                     }
-                                } label: {
+                                } label:
+                                {
                                     Text("Delete")
                                 }
                                 .tint(.red)
@@ -167,8 +188,10 @@ struct ForumView: View {
                 }
                 .listStyle(.sidebar)
                 .tint(.white)
-                .refreshable {
-                    Task {
+                .refreshable
+                {
+                    Task
+                    {
                         await self.getForum()
                     }
                 }
@@ -176,29 +199,37 @@ struct ForumView: View {
             }
         }
         //MARK: 刪除帳號Alert
-        .alert("Are you sure?", isPresented: self.$confirm) {
+        .alert("Are you sure?", isPresented: self.$confirm)
+        {
             Button("Cancel", role: .cancel) {}
             
-            Button("Confirm", role: .destructive) {
-                Task {
+            Button("Confirm", role: .destructive)
+            {
+                Task
+                {
                     await self.deleteUser()
                 }
             }
-        } message: {
+        } message:
+        {
             Text("This action can't be undo.")
         }
         //MARK: 結果Alert
-        .alert(self.result.1, isPresented: self.$result.0) {
+        .alert(self.result.1, isPresented: self.$result.0)
+        {
             Button("Dismiss") {}
         }
         //MARK: Firebase查詢
         .onAppear {
-            if(self.user.account.isEmpty) {
-                withAnimation(.easeInOut.speed(2)) {
+            if(self.user.account.isEmpty)
+            {
+                withAnimation(.easeInOut.speed(2))
+                {
                     self.logIn=false
                 }
             } else {
-                Task {
+                Task
+                {
                     await self.getForum()
                 }
             }
@@ -207,24 +238,29 @@ struct ForumView: View {
         .navigationBarTitleDisplayMode(.large)
         .toolbar {
             //MARK: 帳號操作
-            ToolbarItem(placement: .navigationBarLeading) {
+            ToolbarItem(placement: .navigationBarLeading)
+            {
                 Menu("Account") {
                     Button("Log Out") {
                         self.user.deleteUser()
-                        withAnimation(.easeInOut.speed(2)) {
+                        withAnimation(.easeInOut.speed(2))
+                        {
                             self.logIn=false
                         }
                     }
                     
-                    Button("Delete", role: .destructive) {
+                    Button("Delete", role: .destructive)
+                    {
                         self.confirm.toggle()
                     }
                 }
             }
             
             //MARK: PostTextView
-            ToolbarItem(placement: .navigationBarTrailing) {
-                NavigationLink(destination: PostTextView()) {
+            ToolbarItem(placement: .navigationBarTrailing)
+            {
+                NavigationLink(destination: PostTextView())
+                {
                     Text("Post").bold()
                 }
             }
